pyproject.toml
[project]
name = "projeto01"
version = "0.1.0"
description = ""
authors = [
    {name = "Anwar Hermuche",email = "anwarhermuche2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "projeto01", from = "."}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
pytest-cov = "^6.2.1"
ruff = "^0.12.3"
taskipy = "^1.14.1"
mkdocs = "^1.6.1"
mkdocstrings-python = "^1.16.12"
pygments = "^2.19.2"
mkdocs-material = "^9.6.15"
pymdown-extensions = "^10.16"

[tool.taskipy.tasks]
test = "pytest -v"
lint = "ruff check . && ruff format ."
run = "pytest -v && ruff check . && ruff format . && python -m src.main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
